+++ ts/src/lib/cfg-block.ts
@@ -1,20 +1,21 @@
-import { CaseStyle } from "kryo/case-style";
-import { ArrayType } from "kryo/types/array";
-import { DocumentIoType, DocumentType } from "kryo/types/document";
-import { $CfgAction, CfgAction } from "./cfg-action";
-import { $CfgLabel, CfgLabel } from "./cfg-label";
+import { TaggedUnionType } from "kryo/types/tagged-union";
+import { $CfgEndBlock, CfgEndBlock } from "./cfg-blocks/cfg-end-block";
+import { $CfgReturnBlock, CfgReturnBlock } from "./cfg-blocks/cfg-return-block";
+import { $CfgSimpleBlock, CfgSimpleBlock } from "./cfg-blocks/cfg-simple-block";
+import { $CfgThrowBlock, CfgThrowBlock } from "./cfg-blocks/cfg-throw-block";
 
-export interface CfgBlock {
-  readonly label: CfgLabel;
-  readonly actions: ReadonlyArray<CfgAction>;
-  readonly next?: CfgLabel;
-}
+export type CfgBlock =
+  CfgEndBlock
+  | CfgReturnBlock
+  | CfgSimpleBlock
+  | CfgThrowBlock;
 
-export const $CfgBlock: DocumentIoType<CfgBlock> = new DocumentType<CfgBlock>(() => ({
-  properties: {
-    label: {type: $CfgLabel},
-    actions: {type: new ArrayType({itemType: $CfgAction, maxLength: Infinity})},
-    next: {type: $CfgLabel, optional: true},
-  },
-  changeCase: CaseStyle.SnakeCase,
+export const $CfgBlock: TaggedUnionType<CfgBlock> = new TaggedUnionType<CfgBlock>(() => ({
+  variants: [
+    $CfgEndBlock,
+    $CfgReturnBlock,
+    $CfgSimpleBlock,
+    $CfgThrowBlock,
+  ],
+  tag: "type",
 }));